pyproject.toml
[tool.poetry]
name = "cursovay2"
version = "0.1.0"
description = ""
authors = ["ttt.temaa <artem1005200776@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pytest-cov = "^5.0.0"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
xlrd = "^2.0.1"
pytest = "^8.2.2"
black = "^24.4.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
mypy = "^1.10.1"
black = "^24.4.2"
flake8 = "^7.1.0"
pandas-stubs = "^2.2.2.240603"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119